Move domain format checks out of validators
