+++ src/handlers/deeplinks.ts
@@ -12,11 +12,12 @@ import { analytics } from '@/analytics';
 import { showWalletConnectToast } from '@/components/toasts/WalletConnectToast';
 import { getUniqueId } from '@/entities/assetId';
 import { FiatProviderName } from '@/entities/f2c';
+import { isENSAddressFormat } from '@/features/address/core/domainFormat';
 import { GasSpeed } from '@/features/gas/types/gasSpeed';
 import { useBackendNetworksStore } from '@/features/network/stores/backendNetworksStore';
 import { startSendFromEthereumUrl } from '@/features/transfer/utils/startSendFromEthereumUrl';
 import { pair as pairWalletConnect, setHasPendingDeeplinkPendingRedirect } from '@/features/wallet-connect/services/pair';
-import { checkIsValidAddressOrDomain, isENSAddressFormat } from '@/helpers/validators';
+import { checkIsValidAddressOrDomain } from '@/helpers/validators';
 import { logger } from '@/logger';
 import { type InitialRoute } from '@/navigation/initialRoute';
 import Navigation from '@/navigation/Navigation';
