+++ src/hooks/useImportingWallet.ts
@@ -6,6 +6,7 @@ import { useDispatch } from 'react-redux';
 import { analytics } from '@/analytics';
 import { IS_TEST } from '@/env';
 import { isValidAddress, looksLikeAddress } from '@/features/address/core/address';
+import { isENSAddressFormat, isUnstoppableAddressFormat } from '@/features/address/core/domainFormat';
 import { backupsStore } from '@/features/backup/stores/backupsStore';
 import { PROFILES } from '@/features/config/constants/experimental';
 import { useExperimentalFlag } from '@/features/config/hooks/experimentalHooks';
@@ -14,7 +15,7 @@ import { fetchReverseRecord } from '@/features/ens/utils/handlers';
 import { ChainId } from '@/features/network/types/backendNetworks';
 import { getProvider, isValidBluetoothDeviceId, resolveUnstoppableDomain } from '@/handlers/web3';
 import { WrappedAlert as Alert } from '@/helpers/alert';
-import { isENSAddressFormat, isUnstoppableAddressFormat, isValidWallet } from '@/helpers/validators';
+import { isValidWallet } from '@/helpers/validators';
 import WalletBackupStepTypes from '@/helpers/walletBackupStepTypes';
 import walletBackupTypes from '@/helpers/walletBackupTypes';
 import { WalletLoadingStates } from '@/helpers/walletLoadingStates';
